move DetailFragment proper package
